# Add support for Spark max_by and min_by aggregate functions