… velox/functions/lib/aggregates/MinMaxByAggregatesBase.h
… velox/functions/prestosql/aggregates/MinMaxByAggregates.cpp