Summary:
1. Spark's max_by/min_by updates accumulator when new value is greater than or equal/less than or equal to old value, but Presto is greater than/less than, detailed #4970.
2. Extract common codes of 2-arg version MinMaxByAggregates into velox/functions/lib/aggregates/MinMaxByAggregateBase.h.
3. Add new UT for Spark specific behavior.

Fixed #4970