@mbasmanova 2023-05-30
@Yohahaha I'm not sure I understand why this change is needed. min_by and max_by functions are non-deterministic in the presence of ties, hence, I think current implementation is perfectly valid for Spark, no?