@Yohahaha 2023-05-30
Thanks for reviewing! @mbasmanova 

I think min_by/max_by 's result is related with second column, and found query results correctness problem when testing actually.
Considering below example sql
```
SELECT max_by(x, y) FROM VALUES ('a', 10), ('b', 50), ('c', 50) AS tab(x, y)

Spark result is c
Presto result is b
```
Suppose current accumulator's value is b, next value is c, Spark will update accumulator when if 50 >= 50, but Presto and velox will update when if 50 > 50, that is difference.

Spark's source code https://github.com/apache/spark/blob/52a5d8dfe0d0a9c85a8ac86be9e626d638510736/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/aggregate/MaxByAndMinBy.scala#L104-L109