@mbasmanova 2023-05-30
@Yohahaha In a distributed execution environment it is not possible to guarantee the order of inputs, hence, both b and c are correct results. The only way to ensure order is by using ORDER BY clause (which is not yet supported in Velox).