@Yohahaha 2023-05-30
> In a distributed execution environment it is not possible to guarantee the order of inputs

I agree it.

Let's just talk about function implement details between Spark and Presto, for max_by if input were (b, 1), (c, 1), Presto return b, Spark return c, if input were (c, 1), (b, 1), Presto return c, Spark return b, Presto implement a strictly comparator but Spark not.
Whether shuffle or not, deterministic input get deterministic output right?

This PR won't change current implementations, just reuse common logic and add a specific comparator.

Thanks again, I hope I explained clearly why this change is needed.