@Yohahaha 2023-06-05
Hi @mbasmanova 

Could you give some more comments on this PR? I run a simple agg with shuffle query within Gluten and compared with Vanilla Spark's results, then I find max_by/min_by are really implemented with a  non-strictly comparator, thanks!