@mbasmanova 2023-06-06
@Yohahaha Thank you for clarifying the motivation for the PR. Accidentally, I'm also seeing a bug in Presto's implementation described in https://github.com/facebookincubator/velox/issues/5145 . I wonder if you could help look into it as well.

You are saying that in Spark min_by returns the "last" value associated with the minimum value of "compare", but Velox implementation returns the "first" value. In a distributed execution, without the ORDER BY within the aggregation, it is not possible to guarantee the order of inputs, hence, the definition of "first" and "last" is non-deterministic. At the same time, it seems wasteful to update the accumulator again and again for the same value of "compare". Hence, my preference would be not to do that. 

I understand that the Spark unit test is failing because it runs single-threaded and tests that the implementation returns the "last" value. However, I suspect this unit test only applies to a specific implementation of the function and doesn't verify its required semantics. Hence, I'd ignore the results of that test.