@Yohahaha 2023-06-07
> Accidentally, I'm also seeing a bug in Presto's implementation described in https://github.com/facebookincubator/velox/issues/5145 . I wonder if you could help look into it as well.

@mbasmanova Thanks for mentioned, I provide an optional way to fix that issue.