@Yohahaha 2023-06-07
For this PR, some user in our side require Gluten's behavior must consistent with vanilla Spark, and there should be some cases that input is deterministic I think.

Overall, this PR just want provide a extra comparator which implementation is same with Spark, and this implementation only affect Spark side, hope this reason would make sense for you, thanks! @mbasmanova