@mbasmanova 2023-06-07
> this PR just want provide a extra comparator which implementation is same with Spark

Makes sense. Let's just keep in mind that this behavior will not hold in a general case.