@Yohahaha 2023-07-17
>> this PR just want provide a extra comparator which implementation is same with Spark

> Makes sense. Let's just keep in mind that this behavior will not hold in a general case.

Hi @mbasmanova 

I have verified this change is needed, and extract common logic of max_by/min_by to `functions/lib`, would you help take a look again?