@mbasmanova 2023-07-17
> Extract common codes of MinMaxByAggregates into velox/functions/lib/aggregates/MinMaxByAggregateBase.h, and only affect 2-args version.

Does Spark not support 3-arg version?