@Yohahaha 2023-07-17
> > Extract common codes of MinMaxByAggregates into velox/functions/lib/aggregates/MinMaxByAggregateBase.h, and only affect 2-args version.
> 
> Does Spark not support 3-arg version?

Yes, Spark does not support it, see this doc https://spark.apache.org/docs/latest/api/sql/#max_by