@Yohahaha 2023-07-26
Thanks for your advices, let me try to refactor it.