@Yohahaha 2023-07-28
Hi @kgpai 

I got CI failure due to newly added benchmark, but I thinks this PR should not cause below problem, could you help take a look and give some advices? thanks!

```
  Summary (5.0% threshold):
    Pass: 124 (1 are faster): 
      VectorSlice.cpp/arrayCopy (+10.28%)
    Fail: 1
      VectorFuzzer.cpp/constantInteger (-27.07%)
```