@Yohahaha 2023-07-28
@mbasmanova 
Newly updates:
1. Add template Comparator for MinMaxAggregatesBase, remove MinAggregate/MaxAggregate.
2. Add Spark specific comparator.
3. Add docs.

Could you help review again? thanks!