@Yohahaha 2023-07-28
> Would you double check the performance of the aggregations after these changes?

I try to use SimpleAggregatesBenchmark to check performance, but it seems has not support 2 input args.
Could you give some advices to guide me do benchmark?