@mbasmanova 2023-07-28
@Yohahaha One option is to extend the benchmark. Another option is to run some queries on a cluster. How do you usually evaluate performance?