@Yohahaha 2023-07-28
> One option is to extend the benchmark. Another option is to run some queries on a cluster. 

Let me try with option 1). Thanks!

> How do you usually evaluate performance?

We always run TPC-DS/TPC-H queries in cluster with Spark + Velox for benchmark.