@mbasmanova 2023-07-28
> We always run TPC-DS/TPC-H queries in cluster with Spark + Velox for benchmark.

Got it. These queries do not use min_by/max_by aggregates, do they?