@Yohahaha 2023-07-28
> These queries do not use min_by/max_by aggregates, do they?

Sure, they do not use these functions.