@Yohahaha 2023-07-30
Before this PR
```
============================================================================
[...]gates/benchmarks/SimpleAggregates.cpp     relative  time/iter   iters/s
============================================================================
doRun(max_by_array)                                       228.56ms      4.38
doRun(max_by_norm)                                        313.72ms      3.19
doRun(max_by_hash)                                           1.66s   604.12m
doRun(max_by_array_nulls)                                 227.82ms      4.39
doRun(max_by_norm_nulls)                                  313.84ms      3.19
doRun(max_by_hash_nulls)                                     1.72s   582.86m
----------------------------------------------------------------------------
```
After this PR
```
============================================================================
[...]gates/benchmarks/SimpleAggregates.cpp     relative  time/iter   iters/s
============================================================================
doRun(max_by_array)                                       209.11ms      4.78
doRun(max_by_norm)                                        295.42ms      3.39
doRun(max_by_hash)                                           1.70s   586.92m
doRun(max_by_array_nulls)                                 209.06ms      4.78
doRun(max_by_norm_nulls)                                  294.46ms      3.40
doRun(max_by_hash_nulls)                                     1.63s   613.88m
----------------------------------------------------------------------------
```

@mbasmanova  I use below codes to benchmark, above performance results show after this PR, we have no regression.
```
BENCHMARK_NAMED_PARAM(doRun, max_by_array, "k_array", "max_by(i32, i64)");
BENCHMARK_NAMED_PARAM(doRun, max_by_norm, "k_norm", "max_by(i32, i64)");
BENCHMARK_NAMED_PARAM(doRun, max_by_hash, "k_hash", "max_by(i32, i64)");
BENCHMARK_NAMED_PARAM(doRun, max_by_array_nulls, "k_array", "max_by(i32_halfnull, i64_halfnull)");
BENCHMARK_NAMED_PARAM(doRun, max_by_norm_nulls, "k_norm", "max_by(i32_halfnull, i64_halfnull)");
BENCHMARK_NAMED_PARAM(doRun, max_by_hash_nulls, "k_hash", "max_by(i32_halfnull, i64_halfnull)");
```